add props suggestions for svelte tags
