+++ src/main/java/dev/blachut/svelte/lang/codeInsight/SvelteComponentTagDescriptor.kt
@@ -1,8 +1,11 @@
 package dev.blachut.svelte.lang.codeInsight
 
+import com.intellij.lang.ecmascript6.psi.ES6FromClause
 import com.intellij.lang.javascript.psi.JSElement
+import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.impl.source.xml.XmlDescriptorUtil
+import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.xml.XmlAttribute
 import com.intellij.psi.xml.XmlTag
 import com.intellij.xml.XmlAttributeDescriptor
@@ -10,6 +13,7 @@ import com.intellij.xml.XmlElementDescriptor
 import com.intellij.xml.XmlElementsGroup
 import com.intellij.xml.XmlNSDescriptor
 import com.intellij.xml.impl.schema.AnyXmlAttributeDescriptor
+import dev.blachut.svelte.lang.psi.SvelteFile
 import org.jetbrains.annotations.NonNls
 import java.util.*
 
@@ -32,9 +36,28 @@ class SvelteComponentTagDescriptor(private val myName: String, private val myDec
     override fun getDeclaration(): JSElement = myDeclaration
 
     override fun getAttributesDescriptors(context: XmlTag?): Array<XmlAttributeDescriptor> {
-        // TODO Find props and global Svelte attributes besides slot
-        return knownAttributeDescriptors
-//        return HtmlNSDescriptorImpl.getCommonAttributeDescriptors(context)
+        if (context == null) {
+            return XmlAttributeDescriptor.EMPTY
+        }
+
+        var attributeDescriptors = knownAttributeDescriptors
+
+        // TODO Find global Svelte attributes besides slot and props
+        if (StringUtil.isCapitalized(name)) {
+            val fromClause = PsiTreeUtil.findChildOfType(declaration.parent, ES6FromClause::class.java)
+            val componentReference = fromClause?.resolveReferencedElements()?.first()
+            if (componentReference != null && componentReference is SvelteFile) {
+                val props = ComponentPropsProvider().getComponentProps(
+                        componentReference.viewProvider.virtualFile,
+                        context.project
+                )
+                val propsDescriptors = props?.map { AnyXmlAttributeDescriptor(it) }
+                if (propsDescriptors != null) {
+                    attributeDescriptors += propsDescriptors
+                }
+            }
+        }
+        return attributeDescriptors
     }
 
     override fun getAttributeDescriptor(attribute: XmlAttribute): XmlAttributeDescriptor? {
