+++ src/main/java/dev/blachut/svelte/lang/codeInsight/SvelteTagProvider.kt
@@ -3,10 +3,6 @@ package dev.blachut.svelte.lang.codeInsight
 import com.intellij.codeInsight.completion.PrioritizedLookupElement
 import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
-import com.intellij.lang.html.HTMLLanguage
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.psi.PsiManager
 import com.intellij.psi.html.HtmlTag
 import com.intellij.psi.impl.source.xml.XmlElementDescriptorProvider
 import com.intellij.psi.search.FileTypeIndex
@@ -81,7 +77,7 @@ class SvelteTagProvider : XmlElementDescriptorProvider, XmlTagNameProvider {
         val svelteFiles = FileTypeIndex.getFiles(SvelteFileType.INSTANCE, GlobalSearchScope.allScope(tag.project))
         val reachableComponents = svelteFiles.map {
             val componentName = it.nameWithoutExtension
-            val componentProps = this.getComponentProps(it, tag.project)
+            val componentProps = ComponentPropsProvider().getComponentProps(it, tag.project)
 
             val lookupObject = mapOf("file" to it, "props" to componentProps)
             var lookupElement = LookupElementBuilder.create(lookupObject, componentName)
@@ -99,19 +95,5 @@ class SvelteTagProvider : XmlElementDescriptorProvider, XmlTagNameProvider {
         // TODO Link component documentation
         // TODO Include svelte internal components
         elements.addAll(reachableComponents)
-    }
-
-    private fun getComponentProps(file: VirtualFile, project: Project): List<String?>? {
-        val viewProvider = PsiManager.getInstance(project).findViewProvider(file) ?: return null
-        val psiFile = viewProvider.getPsi(HTMLLanguage.INSTANCE) ?: return null
-
-        val visitor = SvelteScriptVisitor()
-        psiFile.accept(visitor)
-        val jsElement = visitor.jsElement ?: return null
-
-        val propsVisitor = PropsVisitor()
-        jsElement.accept(propsVisitor)
-
-        return propsVisitor.props
     }
 }
